Rename VL to VLBridge, add support for specific buckets
