+++ examples/bridge/histogram/simple.html
@@ -91,7 +91,7 @@
           return Math.round(e);
         };
 
-        const catastroLayer = new AsBridge.VL(
+        const catastroLayer = new AsBridge.VLBridge(
           carto,
           map,
           vizLayer,
