+++ examples/bridge/time-series/simple.html
@@ -87,7 +87,7 @@
           return Math.round(e);
         };
 
-        const catastroLayer = new AsBridge.VL(
+        const catastroLayer = new AsBridge.VLBridge(
           carto,
           map,
           vizLayer,
