+++ packages/bridge/examples/cordoba.html
@@ -88,7 +88,7 @@
           return Math.round(e);
         };
 
-        const bridge = new AsBridge.VL(
+        const bridge = new AsBridge.VLBridge(
           carto,
           map,
           vizLayer,
