+++ packages/components/package.json
@@ -11,7 +11,7 @@
   ],
   "scripts": {
     "build": "stencil build",
-    "dev": "stencil build --dev --watch",
+    "dev": "stencil build --dev --watch --es5",
     "lint:css:fix": "stylelint 'src/components/**/*.scss' --fix",
     "lint:css": "stylelint 'src/components/**/*.scss'",
     "lint:fix": "npm run lint:css:fix",
