+++ packages/components/src/components/as-histogram-widget/arbnbindings.html
@@ -95,7 +95,7 @@
         const latWidget = document.querySelector('#lat');
         const lonWidget = document.querySelector('#lon');
 
-        const listingsLayer = new AsBridge.VL(
+        const listingsLayer = new AsBridge.VLBridge(
           carto,
           map,
           vizLayer
