+++ packages/components/src/components/as-histogram-widget/cordobindings.html
@@ -92,7 +92,7 @@
           return Math.round(e);
         };
 
-        const bridge = new AsBridge.VL(
+        const bridge = new AsBridge.VLBridge(
           carto,
           map,
           vizLayer,
