+++ packages/components/tsconfig.json
@@ -19,7 +19,7 @@
   },
   "include": [
     "src",
-    "types/jsx.d.ts"
+    "types/jsx.d.ts", "../bridge/src/vl/histogram/CategoricalHistogramFilter.ts", "../bridge/src/vl/histogram/NumericalHistogramFilter.ts", "../bridge/src/vl/histogram/BaseHistogramFilter.ts", "../bridge/src/vl/histogram/CategoricalFilter.ts", "../bridge/src/vl/category/CategoryFilter.ts"
   ],
   "exclude": [
     "node_modules"
